Add support for authentication with gitlab CI
Gitlab CI predefine many variables in its environment, among them the
CI hostname and a token that can be used to authenticate with this
server. If we find these variables in the environment add the
credentials to .netrc which in turn allow git and other tools to
access resources found on the CI server.

Signed-off-by: Alban Bedel <alban.bedel@aerq.com>